+++ static/audited_at.json
@@ -24,6 +24,7 @@
   "Primer::Alpha::NavList": "",
   "Primer::Alpha::NavList::Item": "",
   "Primer::Alpha::NavList::Section": "",
+  "Primer::Alpha::OcticonSymbols": "",
   "Primer::Alpha::SegmentedControl": "",
   "Primer::Alpha::SegmentedControl::Item": "",
   "Primer::Alpha::TabNav": "",
