+++ static/statuses.json
@@ -24,6 +24,7 @@
   "Primer::Alpha::NavList": "alpha",
   "Primer::Alpha::NavList::Item": "alpha",
   "Primer::Alpha::NavList::Section": "alpha",
+  "Primer::Alpha::OcticonSymbols": "alpha",
   "Primer::Alpha::SegmentedControl": "alpha",
   "Primer::Alpha::SegmentedControl::Item": "alpha",
   "Primer::Alpha::TabNav": "alpha",
@@ -84,7 +85,7 @@
   "Primer::MenuComponent": "deprecated",
   "Primer::Navigation::TabComponent": "alpha",
   "Primer::OcticonComponent": "deprecated",
-  "Primer::OcticonSymbolsComponent": "alpha",
+  "Primer::OcticonSymbolsComponent": "deprecated",
   "Primer::PopoverComponent": "deprecated",
   "Primer::SpinnerComponent": "beta",
   "Primer::StateComponent": "beta",
