+++ test/components/component_test.rb
@@ -111,6 +111,7 @@ class PrimerComponentTest < Minitest::Test
 
   def test_registered_components
     ignored_components = [
+      "Primer::OcticonSymbolsComponent",
       "Primer::OcticonComponent",
       "Primer::Markdown",
       "Primer::MenuComponent",
